feat(project): add items bindable
This enables users to use and array with values to auto-complete on. When using
this bindable it disables the performing of requests.

Once noticable change is that an empty string will trigger a new search/filter
when the input receives focus.

Contains the necessary documentation which explains what methods can be
overwritten and the newly defined items bindable. Also updated the altered
bindables behavior documentation